9-17-25 Update - Pagination and More
Added pagination to the reviews page. Also added in new fields in the admin add review form for client image and also an image of the job that the review is about. Increased image height on the reviews page to allow more of the image to be seen by potential clients.